refactor: extract error list on its own file
